Combined effects file
