--- BNA_KC_Vehicles/Static/Mortar/CfgComplexEffects.hpp
@@ -1,24 +1,0 @@
-class BNA_KC_ComplexEffects_SmokeShellBlueSmall
-{
-    class SmokeShellBlue
-    {
-        type = "BNA_KC_Cloudlet_SmokeShellBlueSmall";
-        simulation = "particles";
-
-        position[] = {0, 0, 0};
-        intensity = 1;
-        interval = 1;
-    };
-};
-class BNA_KC_ComplexEffects_SmokeShellRedSmall
-{
-    class SmokeShellRed
-    {
-        type = "BNA_KC_Cloudlet_SmokeShellRedSmall";
-        simulation = "particles";
-
-        position[] = {0, 0, 0};
-        intensity = 1;
-        interval = 1;
-    };
-};
