+++ BNA_KC_Vehicles/Static/Mortar/Config.cpp
@@ -1,6 +1,5 @@
 #include "CfgPatches.hpp"
 #include "CfgCloudlets.hpp"
-#include "CfgComplexEffects.hpp"
 #include "CfgFunctions.hpp"
 #include "Macros.hpp"
